requirements.txt
# torch
# torchvision
mlflow
fastapi
uvicorn
pillow